refactor: 추상화 계층에 따라 폴더구조 분리
