+++ src/__tests__/repeat/unit/actions.spec.ts
@@ -1,12 +1,12 @@
-import { getFilteredEvents } from '../../../../utils/eventUtils.ts';
+import { getFilteredEvents } from '../../../utils/eventUtils.ts';
 import {
   expandEventsToNextOccurrences,
   getNextDailyOccurrence,
-} from '../../../../utils/repeatUtils.ts';
-import { getNextWeeklyOccurrence } from '../../../../utils/repeatUtils.ts';
-import { getNextYearlyOccurrence } from '../../../../utils/repeatUtils.ts';
-import { generateInstances } from '../../../../utils/repeatUtils.ts';
-import { makeEvent } from '../../../utils.ts';
+} from '../../../utils/repeat/actions.ts';
+import { getNextWeeklyOccurrence } from '../../../utils/repeat/actions.ts';
+import { getNextYearlyOccurrence } from '../../../utils/repeat/actions.ts';
+import { generateInstances } from '../../../utils/repeat/actions.ts';
+import { makeEvent } from '../../utils.ts';
 
 describe('getNextDailyOccurrence', () => {
   it('기준일보다 같거나 뒤의 첫 daily 발생일을 반환한다', () => {
